FFGraph fixes: * rename IFileSource to ISourceHandle (with changes from SoundBridge). * Improve IOHandler, this can cause regresions. * fix typos and wrong integer casts. * better error handling in FFGraphFormat. * when seeking, avoid skipping the decoded data. * add pts and timebase checks * do not force FFMpeg SWS to have bilinear filtering

+++ src-desktop/ffgraph/converter_audio.cs
@@ -92,7 +92,7 @@ internal unsafe class AudioConverter : IFFGraphFormatConverter {
         NativeMemory.Copy(remaining_buffer, this.sample_buffer, remaining_bytes);
 
         this.sample_buffer_used -= forward_samples;
-        this.seconds += forward_samples / this.sample_rate; // calc again due ceiling
+        this.seconds += (double)forward_samples / this.sample_rate; // calc again due ceiling
     }
 
     public void Dispose() {
